+++ apiserver/paasng/paasng/platform/engine/deploy/bg_wait/wait_deployment.py
@@ -25,8 +25,7 @@ from typing import Dict, List, Optional, Type
 
 from blue_krill.async_utils.poll_task import PollingMetadata, PollingResult, PollingStatus, TaskPoller
 
-from paas_wl.bk_app.processes.processes import PlainProcess
-from paas_wl.bk_app.processes.shim import ProcessManager
+from paas_wl.bk_app.processes.processes import PlainProcess, ProcessManager
 from paasng.platform.applications.models import ModuleEnvironment
 from paasng.platform.engine.models import Deployment
 from paasng.platform.engine.processes.events import ProcEventsProducer
