+++ apiserver/paasng/paasng/platform/engine/views/release.py
@@ -24,7 +24,7 @@ from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 
 from paas_wl.bk_app.applications.api import get_latest_build_id
-from paas_wl.bk_app.processes.shim import ProcessManager
+from paas_wl.bk_app.processes.processes import ProcessManager
 from paasng.accessories.publish.entrance.exposer import env_is_deployed, get_exposed_url
 from paasng.accessories.publish.entrance.preallocated import get_preallocated_url
 from paasng.infras.accounts.permissions.application import application_perm_class
